requirements.txt
pandas==1.4.1
Pillow==9.1.1
requests==2.27.1
streamlit==1.9.2
